+++ src/idtrackerai/video.py
@@ -392,7 +392,8 @@ class Video:
 
     @property
     def blobs_no_gaps_path(self) -> Path:
-        """get the path to save the blob collection after segmentation.
+        """DEPRECATED since v5.2.2
+        get the path to save the blob collection after segmentation.
         It checks that the segmentation has been successfully performed"""
         return self.preprocessing_folder / "list_of_blobs_no_gaps.pickle"